indai-assistant/requirements.txt
SpeechRecognition
pyaudio
pystray
Pillow
pyttsx3
ollama
faster-whisper
pyautogui
requests
sounddevice
numpy
scipy
